Explicitly disable User-Agent header
As discussed in #46, we've been playing "whack-a-mole" with User-Agent
for a while now. It appears that not sending any User-Agent at all is
accepted by myq, at least for now.

This change explicitly disables User-Agent. Note that this is not the
same as not specifying a custom value (which was tried before) - in that
case, aiohttp adds its own default, which apparently got blacklisted at
one point.. This change tells aiohttp to not do that.